Moved [AssetCriticalityNotifier] to Its Own File
class [AssetCriticalityNotifier] has been moved to criticality/asset_criticality_notifier.dart
